feat: remove block building feature entirely
Delete the block building system (buildingTypes.ts, buildingSystem.ts)
and all related Game3D.tsx code — build mode toggle (B key), ghost block
preview, place/remove block callbacks, block collision physics, block
persistence, and the build-mode HUD.

World item persistence (saveWorldItems/loadWorldItems) extracted to the
new lib/worldItemsPersistence.ts so pumpkin/bomb placement is unaffected.
Tests migrated to __tests__/worldItemsPersistence.test.ts. README and
docs updated to reflect the removal.
